+++ tests/testthat/test_get_nldi.R
@@ -94,3 +94,39 @@ test_that("get feature works", {
   expect_equal(f$identifier, "USGS-05428500")
 
 })
+
+test_that("characteristics", {
+  skip_on_cran()
+
+  expect_error(discover_nldi_characteristics(type = "test"), "Type must be one of all, local, total, divergence_routed")
+
+  m <- discover_nldi_characteristics()
+
+  expect_equal(names(m), c("local", "total", "divergence_routed"))
+
+  expect_equal(names(m$local), c("characteristic_id", "characteristic_description", "units", "dataset_label", "dataset_url", "theme_label", "theme_url", "characteristic_type"))
+
+  m <- discover_nldi_characteristics(type = "local")
+
+  expect_equal(names(m$local), c("characteristic_id", "characteristic_description", "units", "dataset_label", "dataset_url", "theme_label", "theme_url", "characteristic_type"))
+
+  m <- discover_nldi_characteristics(type = "total")
+
+  expect_equal(names(m$total), c("characteristic_id", "characteristic_description", "units", "dataset_label", "dataset_url", "theme_label", "theme_url", "characteristic_type"))
+
+  site <- list(featureSource = "nwissite", featureID = "USGS-05429700")
+
+  chars <- get_nldi_characteristics(site)
+
+  expect_equal(names(chars), "local")
+
+  expect_equal(names(chars$local), c("characteristic_id", "characteristic_value", "percent_nodata"))
+
+  chars <- get_nldi_characteristics(site, type = "all")
+
+  expect_equal(names(chars), c("local", "total", "divergence_routed"))
+
+  chars <- get_nldi_characteristics(site, type = "total")
+
+  expect_equal(names(chars), "total")
+})